pyproject.toml
[project]
name = "metafrasis"
version = "0.1.0"
description = "Ancient Greek OCR, Transliteration, and Translation Application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.31.0",
    "pytesseract>=0.3.10",
    "pillow>=10.2.0",
    "opencv-python>=4.9.0",
    "numpy>=2.0.0",
    "transformers>=4.37.0",
    "torch>=2.2.0",
    "sentencepiece>=0.1.99",
    "python-dotenv>=1.0.0",
    "pdf2image>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "matplotlib>=3.8.0",
]

# OCR component dependencies
craft = [
    "scipy>=1.13.0",  # For connected components (supports numpy 2.0+)
]

db = [
    "pyclipper>=1.3.0",  # For polygon operations
]

kraken = [
    "kraken>=5.2.0",  # Kraken OCR library (supports numpy 2.0+)
    "scipy>=1.13.0",  # Required for numpy 2.0+ compatibility
]

# All OCR components
ocr-full = [
    "scipy>=1.13.0",
    "pyclipper>=1.3.0",
    "kraken>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "ml", "native"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests", "native"]
python_files = "test_*.py"
python_functions = "test_*"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_craft: marks tests requiring CRAFT model weights",
    "requires_crnn: marks tests requiring CRNN model weights",
    "requires_tesseract: marks tests requiring Tesseract installation",
    "requires_mps: marks tests requiring Apple Metal Performance Shaders",
    "requires_native: marks tests requiring native extension",
]
